+++ app/helpers/agilibox/all_helpers.rb
@@ -2,13 +2,14 @@ module Agilibox::AllHelpers
   include Agilibox::BootstrapHelper
   include Agilibox::ButtonHelper
   include Agilibox::FiltersHelper
+  include Agilibox::FontAwesomeHelper
   include Agilibox::FormHelper
+  include Agilibox::I18nHelper
   include Agilibox::LinkHelper
   include Agilibox::PaginationHelper
   include Agilibox::RoutesHelper
   include Agilibox::SortingHelper
   include Agilibox::TextHelper
-  include Agilibox::I18nHelper
 
   extend self
 end