+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `boolean_icon` helper
 - Add i18n attributes
 - Add `Agilibox::PluckToHash`
+- Switch to FontAwesome 5 + `icon` helper improvements
 
 ## 1.0.15